requirements.txt
fastapi
uvicorn
numpy
scikit-learn
joblib
torch==2.1.2
psutil
pyspellchecker
--extra-index-url https://download.pytorch.org/whl/cpu
numpy<2
better-profanity>=0.7.0